final commit

+++ src/main/java/com/example/Dlithe/Khetesh/VehicleService.java
@@ -35,7 +35,7 @@ public class VehicleService {
 		return repo.readByMilage(milage);
 	}
 	
-	public List<String> getByCost(int price){
+	public List<String> getByCost(double price){
 		return repo.readByPrice(price);
 	}
